Fix state_changed fired each update fetch by removing force_update
Now the writing state operation is handled in the DTU,
so there is no manual 'if updated' condition to call async_write_ha_state.
The state is always written. But the state_changed only fires if there is any
real change or if entity has the `force_update` flag.

This flag is not needed, as the inclusion of the **bridge sensor** `lastupdated`
attrib ensures that a repeated press changes entity attrs and thus fires the event.