Limit minimum scan_interval to the default one
In practice, the config parsing over `scan_interval: X` in the yaml
is casting to int, so you cannot set 2.5 secs, or 0.25. BUT
manually setting a 0.X value will be taken as scan_interval=0, which is bad.

This avoids that scenario, by limiting to 0.5 secs (value to almost ensure 1Hz updates)